Enable build in GitHub Actions
This adds a workflow to build the website in GitHub with the aim to publish it
to Statichost.eu.

As far as I can tell Statichost.eu can only pull the default branch from a
repository, so this pushes the build to the 'public' branch in the repository,
which needs to be set as the default branch.